:hand: update the degree function (not returning names yet)
