+++ REQUIRE
@@ -3,3 +3,4 @@ StatsBase
 Combinatorics
 Plots
 RecipesBase
+NamedTuples
